Refine rhythm review followups and docs
