+++ src/lib/rhythm/pattern-loader.ts
@@ -1,5 +1,23 @@
 import { sql } from "drizzle-orm";
 import type { SqliteDatabase } from "@/lib/db/client-sqlite";
+import {
+  getTuneTypeLookupCandidates,
+  normalizeTuneTypeName,
+} from "@/lib/rhythm/tune-type-lookup";
+import type {
+  RhythmPatternCandidateScope,
+  RhythmPatternMetadata,
+  RhythmPatternRequest,
+  RhythmPatternType,
+} from "./pattern-types";
+
+export type {
+  RhythmPatternCandidate,
+  RhythmPatternCandidateScope,
+  RhythmPatternMetadata,
+  RhythmPatternRequest,
+  RhythmPatternType,
+} from "./pattern-types";
 
 const DEFAULT_SAMPLE_BASE_URL = (
   import.meta.env.VITE_R2_AUDIO_BASE_URL?.trim() ?? ""
@@ -23,126 +41,12 @@ const DEFAULT_TEMPO_BY_TYPE: Record<string, number> = {
   polka: 120,
 };
 
-const TUNE_TYPE_NAME_ALIASES: Record<string, string> = {
-  air: "air",
-  bdnce: "barn dance",
-  "barn dance": "barn dance",
-  hland: "highland",
-  highland: "highland",
-  hpipe: "hornpipe",
-  hornpipe: "hornpipe",
-  jigd: "jig",
-  jig: "jig",
-  "double jig": "jig",
-  jigsl: "slip jig",
-  "slip jig": "slip jig",
-  sgjig: "jig (single)",
-  "single jig": "jig (single)",
-  "jig (single)": "jig (single)",
-  mzrka: "mazurka",
-  mazurka: "mazurka",
-  piece: "piece",
-  polka: "polka",
-  reel: "reel",
-  sgreel: "reel",
-  schot: "schottische",
-  schottische: "schottische",
-  setd: "set dance",
-  "set dance": "set dance",
-  slide: "slide",
-  song: "song",
-  strath: "strathspey",
-  strathspey: "strathspey",
-  "three-two": "3/2 hornpipe",
-  waltz: "waltz",
-};
-
-const TUNE_TYPE_LOOKUP_VARIANTS: Record<string, readonly string[]> = {
-  hpipe: ["Hpipe", "Hornpipe"],
-  hornpipe: ["Hornpipe", "Hpipe"],
-  jigd: ["JigD", "Jig", "Double Jig"],
-  jig: ["Jig", "JigD", "Double Jig"],
-  jigsl: ["JigSl", "Slip Jig"],
-  "slip jig": ["Slip Jig", "JigSl"],
-  sgjig: ["SgJig", "Jig (Single)", "Single Jig"],
-  "single jig": ["Single Jig", "Jig (Single)", "SgJig"],
-  "jig (single)": ["Jig (Single)", "Single Jig", "SgJig"],
-  sgreel: ["SgReel", "Single Reel", "Reel"],
-  reel: ["Reel", "SgReel", "Single Reel"],
-  setd: ["SetD", "Set Dance"],
-  "set dance": ["Set Dance", "SetD"],
-  strath: ["Strath", "Strathspey"],
-  strathspey: ["Strathspey", "Strath"],
-};
-
-export interface RhythmPatternRequest {
-  genreId?: string | null;
-  genreName?: string | null;
-  tuneTypeName?: string | null;
-  tuneId?: string | null;
-  userId?: string | null;
-  selectedPatternId?: string | null;
-}
-
-export type RhythmPatternType = "seed" | "full_track";
-
-export type RhythmPatternCandidateScope =
-  | "user_tune"
-  | "tune_default"
-  | "user_default"
-  | "system_default"
-  | "system_pattern";
-
-export interface RhythmPatternCandidate {
-  id: string;
-  name: string;
-  scope: RhythmPatternCandidateScope;
-  patternType: RhythmPatternType;
-  sampleKit: string;
-  hasPremiumAudio: boolean;
-}
-
-export interface RhythmPatternMetadata {
-  genreName: string | null;
-  genreId?: string | null;
-  tuneTypeName: string;
-  tuneTypeId?: string | null;
-  rhythmAbc: string;
-  rhythmSignature: string | null;
-  tuneStructure?: string | null;
-  patternType: RhythmPatternType;
-  tempoQpm: number;
-  sampleKit: string;
-  premiumAudioUrl: string | null;
-  premiumAudioTrimMs: number;
-  premiumAudioSource: "database" | null;
-  premiumAudioSourceTempoQpm: number | null;
-  source: "rhythm_patterns" | "tune_type_fallback";
-  selectedPatternId?: string | null;
-  patternCandidates?: RhythmPatternCandidate[];
-}
-
 type PremiumLoopSelection = {
   source: "database";
   url: string;
   sourceTempoQpm: number | null;
   trimMs: number;
 };
-
-function normalizeTuneTypeName(value: string): string {
-  const normalized = value.trim().toLowerCase();
-  return TUNE_TYPE_NAME_ALIASES[normalized] ?? normalized;
-}
-
-function getTuneTypeLookupCandidates(value: string): string[] {
-  const trimmed = value.trim();
-  const normalized = trimmed.toLowerCase();
-  const variants = TUNE_TYPE_LOOKUP_VARIANTS[normalized] ?? [trimmed];
-
-  return Array.from(
-    new Set([trimmed, ...variants].filter((candidate) => candidate.trim()))
-  );
-}
 
 function sanitizeAbcTitle(value: string): string {
   return value.replace(/[\r\n:|[\]]+/g, " ").trim() || "Rhythm";
